Move to log package
