+++ log/doc.go
@@ -1,0 +1,4 @@
+// Log provides a singleton logger for uses in applications where developers
+// prefer not to thread through a logger instance.
+
+package log
