+++ log/singleton.go
@@ -18,24 +18,26 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package zap
+package log
 
-var singletonLogger Logger
+import "github.com/uber-go/zap"
+
+var singletonLogger zap.Logger
 
 func init() {
-	ConfigureStandard(NewJSONEncoder())
+	ConfigureStandard(zap.NewJSONEncoder())
 }
 
 // Standard returns the standard, singleton logger. Although this behavior is discouraged in many settings,
 // many projects use it regardless and we'd prefer a reference implementation to each service doing it differently.
-func Standard() Logger {
+func Standard() zap.Logger {
 	return singletonLogger
 }
 
 // ConfigureStandard configures the singleton logger. Note that this does not provide any
 // synchronization guarantees -- it is up to you to configure your logger before calling Standard()
 // and using it if you want to customize the options.
-func ConfigureStandard(enc Encoder, options ...Option) Logger {
-	singletonLogger = New(enc, options...)
+func ConfigureStandard(enc zap.Encoder, options ...zap.Option) zap.Logger {
+	singletonLogger = zap.New(enc, options...)
 	return singletonLogger
 }
